✨ feat: control by mic
Add a new argument args.auto_mouth_threshold in args.py, which is used to normalize the volume. When this parameter is set to 0, it means that this feature is not used. Personally, 8000 is a good threshold value.

Add a UI to launcher.py to prompt user input.

Use Pyaudio in main.py to obtain audio input stream and adjust the mouth size.

Note: This feature is written towards the end of main.py, so it overrides the lip-syncing feature's mouth size.

PR by PipeYume at yuyuyzl/EasyVtuber